+++ code/utils/xrLC/xrLC.vcxproj.filters
@@ -270,12 +270,12 @@
     <ClCompile Include="net.cpp">
       <Filter>%2a%2a%2a COMPILER %2a%2a%2a\net</Filter>
     </ClCompile>
-    <ClCompile Include="..\common\NvMender2003\NVMeshMender3.cpp">
-      <Filter>External\NV_Mender2003</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\engine.vc2008\xrEngine\xrLoadSurface.cpp" />
     <ClCompile Include="..\..\SDK\include\hxgrid\Interface\hxGridInterface.cpp">
       <Filter>%2a%2a%2a COMPILER %2a%2a%2a\net</Filter>
+    </ClCompile>
+    <ClCompile Include="..\common\NvMender2003\NVMeshMender4.cpp">
+      <Filter>External\NV_Mender2003</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
